requirements.txt
streamlit
openai
python-dotenv
langchain-core
PyPDF2
